1.1 trying to figure out my error

+++ more_malloc_free/1-string_nconcat.c
@@ -9,7 +9,7 @@
 char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
 	char *new;
-	size_t len = n, index;
+	unsigned int len = n, index;
 
 	if (!s1)
 	{
